Set NoiseRemoval to "false" by default as it doesn't work
NoiseRemoval uses a unidirectional filter which shifts the data, so it removes intervals ~20ms after the supposed "noise" it detected (which is shorter than 20ms). I am also not convinced that peaks in high gamma (I got peak correlation with 70 Hz wavelet power) are necessarily noise to be removed, so I recommend this option not be used at all.

The whitening step in Kilosort should already be good enough to ignore shared noise between channels.