Add an option to DetectSWR to manually adjust the clusters (super useful)
If "check" is set to true (default is false, so it shouldn't change current function behavior), then the function would plot the k-means separation it would normally do and the user should validate it. If the user finds that the separation is terrible (for example, due to bad sharp waves or noisy outliers), they are free to adjust it manually (make new idx1 and idx2 logicals while in debug mode) before writing "dbcont" and letting the function continue with the adjusted clusters.