change default ripple detection window from 200ms to 20ms
DetectSWR first makes a list of candidate events (based on candidate sharp wave) and finds the ripple value associated with each of these candidate events. The ripple value assigned to each event is the highest ripple value within a window of "WindowSize"  around the candidate event.

With "WindowSize" being 200ms by default, in a great number of cases the value of a completely separate event is taken. Candidate events (most of which are not actually ripples) may rapidly follow each other (median distance between 2 candidate events in one example session was ~150ms). Taking the peak ripple value 100ms before/after the moment of the candidate event means that very often our estimate of the "ripple value" of a bad candidate event is estimated as good, just because the bad event occurs close to an actual ripple. Paradoxically, if the actual ripple follows such a bad candidate event within "minIsi" distance, it appears that then the true ripple itself would be rejected, while the bad event is retained.

To avoid this, the window for estimating the ripple power of a candidate event should be drastically reduced. The reason such a window exists at all is to capture cases when the ripple is not visible at the trough of the sharp wave but follows shortly after. I have tested the algorithm with a window of 20ms and I am much happier with the results, but ymmv if the temporal distance between a ripple and a sharp wave is larger in your dataset, in which case you should increase it (but try to keep it as low as possible).