Clarify what the default option "useSPW" means (misleading). I recommend you call DetectSWR with useSPW=false.
Without any explanation, one may erroneously think that "useSPW" indicates whether or not the sharp wave should be taken into account in DetectSWR. However, the sharp wave is always taken into account, and the variable "useSPW" instead indicates that all events without a significant sharp wave will be removed.

Moreover, DetectSWR used the sharp wave only to determine the boundaries of the detected events, even if "useSPW" was set to false. In practice, this means that beautiful ripple events without a sharp wave would be removed at a later stage (due to other duration conditions). This has now been changed, so it is only the case when "useSPW" is true.

In many sessions, one may observe beautiful ripples with and without a prominent sharp wave. If "useSPW" is set to false, then these events would still be detected, taking into account the sharp wave (and likely requiring higher ripple power when a sharp wave is not prominent).