I had recently reduced this parameter to 20ms, but I'm reverting it to a more intermediate value
This is the temporal distance we should expect a ripple and a sharp wave to be separated by.

Before, it was 200ms, which was way too high and we were erroneously admitting noisy events because they were temporally close to real events. I recently reduced it to 20ms (so 10ms before and 10ms after the sharp wave), which appears to be too strict.

For OJR34, the visible sharp-wave tends to follow the ripple by 13ms, so I'm increasing it to 20ms (20ms before+20ms after, i.e. WinSize=40ms) to capture events in this kind of recording.