Relax ripple duration criteria
DetectSWR finds a significant sharp wave interval and a significant ripple interval. The function behavior was that if the significant ripple interval is not above a certain minDurRP threshold, then the event was removed. This is undesirable because in some cases, an event might be characterized by a prominent sharp wave and a short (if any) significant ripple interval. In sessions where a considerable proportion of detected events are in this category, their removal is undesirable.

This changes proposes that events are removed only when the significant sharp wave interval as well as the significant ripple interval are both too short.

The change in line 1122 is just to fix a minor oversight (according to the previous code, the first ripple would get automatically removed for no reason).