Update getRipSpikes.m
Changed input structure to force 'spikes' and 'events' to be something (if empty will spit an empty var). In this way, the script can be used for other event types and not just for ripples (previously if empty or no input vars were given it was considering the whole ripple var from the path). Let me know if this is a hustle for anyone or breaks something along the way..