A completely new implementation fixing multiple issues
There were multiple issues with the KM version of Noise remover.

It was finding intervals with high gamma power instead of high-frequency noise. It was using a unidirectional filter (so removing periods *after* the detected 'noise' intervals). It was using a fixed factor of 8 to convert from lfp samples to dat-file samples (when in our files, dat files are typically 20kHz vs the 1.25kHz of the lfp file, which makes a factor of 16).

I did a complete rewrite, using the same idea of detecting lfp noise and interpolating the noisy periods in the .dat file. Now we detect 300 Hz+ noisy periods, select outliers using Tukey's fences, expand them by [-2.5 +2.5] ms and merge periods within 5ms of each other. Then we remove all the .dat-file timestamps within these noisy intervals, interpolating from the signal around them. The noise intervals are saved for posterity in noiseIntervals.events.mat.

I have tested this on a noisy session and I am happy with the results.