Restoring the GNU GPL license notice. Removing this is a GPL violation.
All of the FMAT functions include the GNU GPL notice. Basically, a GPL license is a "copyleft" license, which enforces libre access to the work and all derivatives. You can freely use and modify the code *as long as your releases remain free software*. Code shared on github without any license is not really free software.

Since we are distributing the code (as it's readily available in our public github), every time you modify and build on code that is under the GNU GPL licence, you need to have your version (if you're distributing it) under that license as well. This prevents the situation where someone shares the code with some company (e.g. by uploading it to FileExchange on Mathworks, which does not allow such licenses) that then decides to copyright it from themselves (e.g. use it in a future toolbox) and suing people who try to build on it in the future.

I always include the GPL license in my code (except when I forget :D).