Add an "eeg" option to DetectSWR and a function to create the EEG
In our data, I find that ripple detection is much better if I were to first treat the lfp by subtracting the mean of all (standard practice in EEG data). In practice, this "cleans" the lfp from random fluctuations that may be mistaken for ripples.

Instead of replacing the raw lfp file (which can be useful to keep to e.g. estimate EMG from correlated fluctuations which would be lost in this new "eeg" version) we can save a different file (.eeg) and ask DetectSWR to use that file (rather than the raw .lfp)