allow computing EMG on a single-shank animal
Previous behavior of this function was to take one channel from each shank and correlate the selected channels. The problem with this is that in an animal in which a single shank is recorded, the EMG would be a series of NaNs.

Solution: in case of a single shank, pick multiple (5) random channels instead of a single one. This allows the correlation step to take place.