improve speed and fix bug
FindClosest(locsB,temp2,'higher') is much faster than a loop.

On line 224, there was a check if all the analog values within a detected pulse are sufficiently high (40% of threshold). The problem was that the threshold was computed based on the absolute values, while the val{jj} contained values with the baseline subtracted (pul{jj} - offset). Therefore, this signal should be compared to an equivalent threshold (baseline subtracted).