Raly fix of the occasional error at the end of creating the LFP file
The final chunk is often not complete. There is a code below dealing with the remainder (line 205 and down) which is meant for such cases. The solution is therefore to decrease our expectations for the number of complete chunks.

I have not attempted to find why the number of complete chunks is lower than (nBytes/(nbChan*sizeInBytes*chunksize)). It might be because we're loading "nbChan*(chunksize+2*ntbuff)" to actually load the file, so for large values of ntbuff we may go over an entire chuink. Regardless, expecting one chunk less and dealing with a slightly larger remainder should fix the issue.