add a faster version of ConsolidateIntervals
As mentioned in the function description, this version makes more assumptions than ConsolidateIntervals, so it's better to keep ConsolidateIntervals for the general case. However, there are many cases where these assumptions are met (e.g. merging non-overlapping noisy intervals within 0.1s of each other), where ConsolidateIntervals is incredibly slow. This is when it's good to use this one to save time.