add check if signal appears to have pulses, and abort if not
I added a threshold for how likely it is that the signal contains pulses. This is because if the signal does not actually contain pulses (but only noise), the pulse detection algorithm will take a very long time (wasted).

The idea behind this check is that if real pulses are present in the signal, the signal is bi-modal (stable high value and stable low values). This bimodality can be captured by the "effectiveness metric" (e.g. for Otsu thresholding). Usually noise sessions have a value of <0.08 and sessions with pulses have a value around >0.97, but I set a somewhat arbitrary threshold of 0.3, because I found example sessions in which noise could reach ~0.2, and in a long session with few (real) pulses, it is possible for it to be as low as 0.5.