change the way errors are computed for noncircular data
While previously, the matrix would be circularly shifted to the current position, this fix makes the shift non-circular, so the matrix is appended with nans.

Example: Before if the probability matrix was
estimations = [0.2;0;0.1;0.6;0.1]
and the correct position is 4, then the (circular) shift would result in an error matrix
errors = [0; 0.01; 0.6; 0.1; 0.2];

which is not ideal in the linear case, since bin 1 (p=0.2) and 5 (p=0.1) are not close. After this edit, for linear data, the error matrix would be:

errors = [0; 0.01; 0.6; 0.1; nan];